refactor: further refactoring of xlsx code
