+++ src/pages-helpers/curriculum/xlsx/builders/buildNcCriteriaText.ts
@@ -1,0 +1,22 @@
+import { cartesianToExcelCoords } from "@ooxml-tools/units";
+
+import { cdata, safeXml } from "../../docx/xml";
+
+export function buildNcCriteriaText<T extends Record<string, string>>(
+  cellStyleIndexMap: T,
+  nationalCurricText: string,
+  x: number,
+  y: number,
+) {
+  return safeXml`
+    <c
+      r="${cartesianToExcelCoords([x, y])}"
+      t="inlineStr"
+      s="${cellStyleIndexMap.temp4!}"
+    >
+      <is>
+        <t xml:space="preserve">${cdata(`\n${nationalCurricText}\n`)}</t>
+      </is>
+    </c>
+  `;
+}
