+++ src/pages-helpers/curriculum/xlsx/builders/buildStyles.ts
@@ -1,0 +1,236 @@
+import { safeXml } from "../../docx/xml";
+import { createXmlIndexMap } from "../helper";
+
+export function buildStyle() {
+  const { xml: fontsXml, indexMap: fontIndexMap } = createXmlIndexMap({
+    arial12: safeXml`
+      <font>
+        <sz val="12" />
+        <color theme="1" />
+        <name val="Arial" />
+        <family val="2" />
+      </font>
+    `,
+    arialBold22: safeXml`
+      <font>
+        <sz val="22" />
+        <color rgb="FFFFFFFF" />
+        <name val="Arial Bold" />
+        <family val="2" />
+      </font>
+    `,
+    arialBold24: safeXml`
+      <font>
+        <sz val="24" />
+        <color rgb="FF000000" />
+        <name val="Arial Bold" />
+        <family val="2" />
+      </font>
+    `,
+    arialBold14: safeXml`
+      <font>
+        <sz val="14" />
+        <color rgb="FF000000" />
+        <name val="Arial Bold" />
+        <family val="2" />
+      </font>
+    `,
+  });
+
+  const { xml: fillsXml, indexMap: fillIndexMap } = createXmlIndexMap({
+    none: safeXml`
+      <fill>
+        <patternFill patternType="none" />
+      </fill>
+    `,
+    temp1: safeXml`
+      <fill>
+        <patternFill patternType="gray125" />
+      </fill>
+    `,
+    temp2: safeXml`
+      <fill>
+        <patternFill patternType="solid">
+          <fgColor rgb="FFF2F2F2" />
+          <bgColor rgb="FFC6EFCE" />
+        </patternFill>
+      </fill>
+    `,
+    temp3: safeXml`
+      <fill>
+        <patternFill patternType="solid">
+          <fgColor rgb="FFC8F2C2" />
+          <bgColor rgb="FFC6EFCE" />
+        </patternFill>
+      </fill>
+    `,
+    temp4: safeXml`
+      <fill>
+        <patternFill patternType="solid">
+          <fgColor rgb="FFEBFCEB" />
+          <bgColor rgb="FFC6EFCE" />
+        </patternFill>
+      </fill>
+    `,
+    temp5: safeXml`
+      <fill>
+        <patternFill patternType="solid">
+          <fgColor rgb="FF000000" />
+          <bgColor rgb="FFFFFFFF" />
+        </patternFill>
+      </fill>
+    `,
+  });
+
+  const { xml: bordersXml, indexMap: borderIndexMap } = createXmlIndexMap({
+    none: safeXml`
+      <border>
+        <left />
+        <right />
+        <top />
+        <bottom />
+        <diagonal />
+      </border>
+    `,
+    lightGray: safeXml`
+      <border>
+        <left style="thin">
+          <color rgb="FFD4D4D4" />
+        </left>
+        <right style="thin">
+          <color rgb="FFD4D4D4" />
+        </right>
+        <top style="thin">
+          <color rgb="FFD4D4D4" />
+        </top>
+        <bottom style="thin">
+          <color rgb="FFD4D4D4" />
+        </bottom>
+        <diagonal />
+      </border>
+    `,
+  });
+
+  const { xml: cellStylesXml, indexMap: cellStyleIndexMap } = createXmlIndexMap(
+    {
+      temp0: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arial12}"
+          fillId="${fillIndexMap.none}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        />
+      `,
+      temp1: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arialBold24}"
+          fillId="${fillIndexMap.temp2}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        >
+          <alignment wrapText="1" vertical="top" />
+        </xf>
+      `,
+      temp2: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arialBold14}"
+          fillId="${fillIndexMap.temp3}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        >
+          <alignment wrapText="1" vertical="top" />
+        </xf>
+      `,
+      temp3: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arialBold14}"
+          fillId="${fillIndexMap.temp3}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        >
+          <alignment wrapText="1" vertical="top" />
+        </xf>
+      `,
+      temp4: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arial12}"
+          fillId="${fillIndexMap.temp4}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        >
+          <alignment wrapText="1" vertical="center" />
+        </xf>
+      `,
+      temp5: safeXml`
+        <xf
+          numFmtId="0"
+          fontId="${fontIndexMap.arialBold22}"
+          fillId="${fillIndexMap.temp5}"
+          borderId="${borderIndexMap.lightGray}"
+          xfId="0"
+          applyBorder="0"
+        >
+          <alignment wrapText="1" horizontal="center" vertical="center" />
+        </xf>
+      `,
+    },
+  );
+
+  const styleXml = safeXml`
+    <?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
+    <styleSheet
+      xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      mc:Ignorable="x14ac x16r2 xr"
+      xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac"
+      xmlns:x16r2="http://schemas.microsoft.com/office/spreadsheetml/2015/02/main"
+      xmlns:xr="http://schemas.microsoft.com/office/spreadsheetml/2014/revision"
+    >
+      <fonts count="1" x14ac:knownFonts="1">${fontsXml}</fonts>
+      <fills count="3">${fillsXml}</fills>
+      <borders count="1">${bordersXml}</borders>
+      <cellStyleXfs count="1">
+        <xf numFmtId="0" fontId="0" fillId="0" borderId="1" />
+      </cellStyleXfs>
+      <cellXfs count="1">${cellStylesXml}</cellXfs>
+      <cellStyles count="1">
+        <cellStyle name="Normal" xfId="0" builtinId="0" />
+      </cellStyles>
+      <dxfs count="0" />
+      <tableStyles
+        count="0"
+        defaultTableStyle="TableStyleMedium2"
+        defaultPivotStyle="PivotStyleLight16"
+      />
+      <extLst>
+        <ext
+          uri="{EB79DEF2-80B8-43e5-95BD-54CBDDF9020C}"
+          xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"
+        >
+          <x14:slicerStyles defaultSlicerStyle="SlicerStyleLight1" />
+        </ext>
+        <ext
+          uri="{9260A510-F301-46a8-8635-F512D64BE5F5}"
+          xmlns:x15="http://schemas.microsoft.com/office/spreadsheetml/2010/11/main"
+        >
+          <x15:timelineStyles defaultTimelineStyle="TimeSlicerStyleLight1" />
+        </ext>
+      </extLst>
+    </styleSheet>
+  `.trim();
+
+  return {
+    styleXml,
+    cellStyleIndexMap,
+  };
+}
