+++ src/pages-helpers/curriculum/xlsx/builders/buildTickCell.ts
@@ -1,0 +1,22 @@
+import { cartesianToExcelCoords } from "@ooxml-tools/units";
+
+import { cdata, safeXml } from "../../docx/xml";
+
+export function buildTickCell<T extends Record<string, string>>(
+  cellStyleIndexMap: T,
+  hasNcCriteria: boolean,
+  x: number,
+  y: number,
+) {
+  return safeXml`
+    <c
+      r="${cartesianToExcelCoords([x, y])}"
+      t="inlineStr"
+      s="${hasNcCriteria ? cellStyleIndexMap.temp5! : cellStyleIndexMap.temp0!}"
+    >
+      <is>
+        <t>${cdata(hasNcCriteria ? "✓" : "")}</t>
+      </is>
+    </c>
+  `;
+}
