+++ src/pages-helpers/curriculum/xlsx/helper.ts
@@ -1,3 +1,5 @@
+import { JSZipCached } from "../docx/docx";
+
 export function createXmlIndexMap<T extends Record<string, string>>(def: T) {
   const indexMap: Record<keyof typeof def, string> = {} as Record<
     keyof typeof def,
@@ -14,4 +16,15 @@ export function createXmlIndexMap<T extends Record<string, string>>(def: T) {
     xml,
     indexMap,
   };
+}
+
+export function addOrUpdateSheet(
+  zip: JSZipCached,
+  sheetNumber: number,
+  xml: string,
+) {
+  zip.writeString(
+    `xl/worksheets/sheet${sheetNumber}.xml`,
+    `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>${xml}`,
+  );
 }
